+++ packages/for-ui/src/modal/ModalHeader.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from 'react';
 
-import { Typography } from '../typography';
+import { Text } from '../text';
 
 type Props = {
   /** The contents of the Dialog. */
@@ -10,9 +10,9 @@ type Props = {
 export const ModalHeader: React.FC<Props> = ({ children }) => {
   return (
     <div className="border-shade-light-default border-b p-4">
-      <Typography variant="p" bold className="text-r font-bold">
+      <Text as="p" className="text-shade-dark-default font-bold">
         {children}
-      </Typography>
+      </Text>
     </div>
   );
 };