+++ packages/for-ui/src/text/Text.stories.tsx
@@ -1,0 +1,14 @@
+import { Meta } from '@storybook/react/types-6-0';
+import { Text } from './Text';
+
+export default {
+  title: 'General / Text',
+  component: Text,
+} as Meta;
+
+export const Base = {
+  args: {
+    children: '「For Design System」で叶えるデザインシステムの夢。',
+    as: 'span',
+  },
+};