+++ packages/for-ui/tailwind.config.base.js
@@ -102,11 +102,6 @@ const colors = {
       light: {
         default: 'var(--primary-text-light-default)',
       },
-      white: {
-        default: 'var(--primary-text-white-default)',
-        hover: 'var(--primary-text-white-hover)',
-        disabled: 'var(--primary-text-white-disabled)',
-      },
     },
 
     background: {
@@ -347,6 +342,28 @@ const fontSizes = {
   ],
 };
 
+const borderColors = {
+  transparent: 'transparent',
+  shade: colors.shade.border,
+  primary: colors.primary.border,
+  secondary: colors.secondary.border,
+  negative: colors.negative.border,
+  positive: colors.positive.border,
+  notice: colors.notice.border,
+  informative: colors.notice.border,
+};
+
+const iconColors = {
+  transparent: 'transparent',
+  shade: colors.shade.icon,
+  primary: colors.primary.icon,
+  secondary: colors.secondary.icon,
+  positive: colors.positive.icon,
+  notice: colors.notice.icon,
+  informative: colors.notice.icon,
+  negative: colors.negative.icon,
+};
+
 module.exports = {
   content: [
     './src/**/*.{js,ts,jsx,tsx}',
@@ -396,26 +413,10 @@ module.exports = {
       notice: colors.notice.background,
       informative: colors.notice.background,
     },
-    borderColor: {
-      transparent: 'transparent',
-      shade: colors.shade.border,
-      primary: colors.primary.border,
-      secondary: colors.secondary.border,
-      negative: colors.negative.border,
-      positive: colors.positive.border,
-      notice: colors.notice.border,
-      informative: colors.notice.border,
-    },
-    iconColor: {
-      transparent: 'transparent',
-      shade: colors.shade.icon,
-      primary: colors.primary.icon,
-      secondary: colors.secondary.icon,
-      positive: colors.positive.icon,
-      notice: colors.notice.icon,
-      informative: colors.notice.icon,
-      negative: colors.negative.icon,
-    },
+    borderColor: borderColors,
+    ringColor: borderColors,
+    iconColor: iconColors,
+    fillColor: iconColors,
     fontSize: {
       xs: fontSizes.s,
       sm: fontSizes.s,
@@ -510,6 +511,7 @@ module.exports = {
       },
       fontFamily: {
         sans: ['YakuHanJP', 'Inter', 'Noto Sans JP', ...defaultTheme.fontFamily.sans],
+        mono: ['Fira Code', 'Noto Sans JP', ...defaultTheme.fontFamily.mono],
       },
       gridTemplateColumns: {
         'auto-1': 'repeat(auto-fill, minmax(80px, 1fr))',