+++ packages/for-ui/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @3design/3design-ui
 
+## 1.0.7
+
+### Patch Changes
+
+- [#921](https://github.com/4-design/for-ui/pull/921) [`b7655a8`](https://github.com/4-design/for-ui/commit/b7655a8d6d5ca9b3b7fcd6c845850146c136dde1) Thanks [@Qs-F](https://github.com/Qs-F)! - fix: MDX で書かれた Storybook が崩れていたのを修正
+
+- [#929](https://github.com/4-design/for-ui/pull/929) [`1518d83`](https://github.com/4-design/for-ui/commit/1518d8325f8b1f635d1b34d8a2f5efe2c5104464) Thanks [@Qs-F](https://github.com/Qs-F)! - fix: 存在しないカラートークンを削除
+
+- [#884](https://github.com/4-design/for-ui/pull/884) [`7889f97`](https://github.com/4-design/for-ui/commit/7889f97c01a4df82c2f6fd855393277e197422dd) Thanks [@renovate](https://github.com/apps/renovate)! - chore(deps): update all non-major dependencies
+
+- [#891](https://github.com/4-design/for-ui/pull/891) [`f33803f`](https://github.com/4-design/for-ui/commit/f33803f1eb0c449477838f7252bfaa8556d39ed3) Thanks [@Qs-F](https://github.com/Qs-F)! - fix(Text): props から`variant`とデフォルトのテキスト色を削除
+
+- [#930](https://github.com/4-design/for-ui/pull/930) [`12f920d`](https://github.com/4-design/for-ui/commit/12f920dcebca696aa93f38f8b2e556f585ab2141) Thanks [@Qs-F](https://github.com/Qs-F)! - fix(Text): Figma との対応関係を追加
+
+- [#911](https://github.com/4-design/for-ui/pull/911) [`de18667`](https://github.com/4-design/for-ui/commit/de186674044c9ee1ae638187148a57eb89b370cc) Thanks [@Qs-F](https://github.com/Qs-F)! - fix(Table): Row Component が `className` props を受け取れるように修正
+
 ## 1.0.6
 
 ### Patch Changes